requirements.txt
requests
python-dotenv
pyrebase4
pypdf2
sentence-transformers
faiss-cpu
spacy
flask
flask-restx
google-generativeai

